refectored register dialog into another component
