+++ src/components/EventComponent/EventInfoPreview.tsx
@@ -1,28 +1,8 @@
 "use client";
 
 import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-  DialogFooter,
-  DialogClose,
-} from "@/components/ui/dialog";
-import {
-  Calendar,
-  Clock,
-  MapPin,
-  Mail,
-  Text,
-  Image as ImageIcon,
-  Users,
-  CalendarClock,
-  X,
-  Check,
-  Download,
-} from "lucide-react";
+import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger, DialogFooter } from "@/components/ui/dialog";
+import { Calendar, Clock, MapPin, Mail, Text, Image as ImageIcon, Users, CalendarClock } from "lucide-react";
 import Image from "next/image";
 import { useState } from "react";
 import {
@@ -38,8 +18,7 @@ import {
 import { useToast } from "@/hooks/use-toast";
 import { useUser } from "@clerk/nextjs";
 import { useRouter } from "next/navigation";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Input } from "@/components/ui/input";
+import { EventRegisterPreview } from "./EventRegisterPreview";
 
 interface EventInfoProps {
   id: string;
@@ -56,160 +35,6 @@ interface EventInfoProps {
   userRegistered?: boolean;
   onDelete?: (eventId: string) => void;
 }
-
-interface AttendeeInfo {
-  name: string;
-  type: "adult" | "child";
-}
-
-interface RegisterDialogProps {
-  isOpen: boolean;
-  onOpenChange: (open: boolean) => void;
-  eventInfo: {
-    title: string;
-    date: string;
-    time: string;
-    location: string;
-    contactEmail: string;
-    waiverDueDate: string;
-  };
-  userInfo: {
-    name: string;
-    family: { name: string }[];
-  };
-  onConfirm: () => void;
-}
-
-const RegisterDialog = ({ isOpen, onOpenChange, eventInfo, userInfo, onConfirm }: RegisterDialogProps) => {
-  const [selectedAttendees, setSelectedAttendees] = useState<string[]>([]);
-  const [waiverEmail, setWaiverEmail] = useState("");
-  const [waiverSigned, setWaiverSigned] = useState(false);
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onOpenChange}>
-      <DialogContent className="w-[60vw] min-w-[600px] p-8">
-        <DialogHeader>
-          <div className="flex items-center justify-between">
-            <DialogTitle className="text-3xl font-bold">Register for: {eventInfo.title}</DialogTitle>
-          </div>
-
-          <div className="mt-4 flex gap-6 text-sm text-gray-600">
-            <div className="flex items-center gap-2">
-              <Calendar className="h-4 w-4" />
-              {eventInfo.date}
-            </div>
-            <div className="flex items-center gap-2">
-              <Clock className="h-4 w-4" />
-              {eventInfo.time}
-            </div>
-            <div className="flex items-center gap-2">
-              <MapPin className="h-4 w-4" />
-              {eventInfo.location}
-            </div>
-            <div className="flex items-center gap-2">
-              <Mail className="h-4 w-4" />
-              {eventInfo.contactEmail}
-            </div>
-          </div>
-        </DialogHeader>
-
-        <div className="mt-6 grid grid-cols-2 gap-12">
-          <div className="w-full">
-            <h3 className="mb-4 font-semibold">Who&apos;s attending?</h3>
-            <div className="space-y-3">
-              <div className="flex items-center space-x-3">
-                <Checkbox
-                  id={userInfo.name}
-                  checked={selectedAttendees.includes(userInfo.name)}
-                  onCheckedChange={(checked) => {
-                    setSelectedAttendees((prev) =>
-                      checked ? [...prev, userInfo.name] : prev.filter((name) => name !== userInfo.name),
-                    );
-                  }}
-                />
-                <label
-                  htmlFor={userInfo.name}
-                  className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-                >
-                  {userInfo.name}
-                </label>
-              </div>
-              {userInfo.family.map((member) => (
-                <div key={member.name} className="flex items-center space-x-3">
-                  <Checkbox
-                    id={member.name}
-                    checked={selectedAttendees.includes(member.name)}
-                    onCheckedChange={(checked) => {
-                      setSelectedAttendees((prev) =>
-                        checked ? [...prev, member.name] : prev.filter((name) => name !== member.name),
-                      );
-                    }}
-                  />
-                  <label
-                    htmlFor={member.name}
-                    className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-                  >
-                    {member.name}
-                  </label>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          <div>
-            <h3 className="mb-4 font-semibold">Waiver (Due {eventInfo.waiverDueDate})</h3>
-            <p className="mb-4 text-sm text-gray-600">
-              To attend this event, please sign this waiver first. You can sign it digitally or print, sign, and upload
-              it.
-            </p>
-
-            <div className="space-y-4">
-              <Button variant="secondary" className="w-full" onClick={() => window.open("/waiver.pdf", "_blank")}>
-                <Download className="mr-2 h-4 w-4" />
-                Download {eventInfo.title} Waiver
-              </Button>
-
-              <div className="text-center text-sm text-gray-600">or</div>
-
-              <div>
-                <label className="text-sm text-gray-600">Send waiver to this email (for printing)</label>
-                <Input
-                  placeholder="e.g. jameshudson345@gmail.com"
-                  value={waiverEmail}
-                  onChange={(e) => setWaiverEmail(e.target.value)}
-                  className="mt-1"
-                />
-              </div>
-
-              {waiverSigned ? (
-                <Button variant="outline" className="w-full bg-green-500 text-white hover:bg-green-600">
-                  <Check className="mr-2 h-4 w-4" />
-                  Signed on {/* Add date here */}
-                  <span className="ml-2 text-sm">Click here to view</span>
-                </Button>
-              ) : (
-                <Button className="w-full bg-[#488644] text-white hover:bg-[#3a6d37]">Click here to sign waiver</Button>
-              )}
-            </div>
-          </div>
-        </div>
-
-        <DialogFooter className="mt-6">
-          {!waiverSigned && (
-            <Button className="mx-auto w-2/5 bg-[#488644] text-white hover:bg-[#3a6d37]" onClick={onConfirm}>
-              Register for Event
-            </Button>
-          )}
-          {waiverSigned && (
-            <Button className="mx-auto w-2/5 bg-[#488644] text-white hover:bg-[#3a6d37]" onClick={onConfirm}>
-              Sign and return to calendar
-            </Button>
-          )}
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
 
 export function EventInfoPreview({
   id,
@@ -457,7 +282,7 @@ export function EventInfoPreview({
           </div>
           <DialogFooter className="flex justify-between">
             <Button
-              className={`text-white ${userRegistered ? "cursor-not-allowed bg-gray-400" : "bg-green-500 hover:bg-green-600"}`}
+              className={`text-white ${userRegistered ? "cursor-not-allowed bg-gray-400" : "bg-[#488644] text-white hover:bg-[#3a6d37]"}`}
               onClick={handleOpenRegisterDialog}
               disabled={isRegisterDisabled}
             >
@@ -505,7 +330,7 @@ export function EventInfoPreview({
         </AlertDialogContent>
       </AlertDialog>
 
-      <RegisterDialog
+      <EventRegisterPreview
         isOpen={isRegisterDialogOpen}
         onOpenChange={setIsRegisterDialogOpen}
         eventInfo={{
